Issue 6306: SLTS - avoid deadlock in truncate (#6307)
In case of system segments, do not attempt to delete block read index entries. Add debug logs for openRead, openWrite and getStreamSegmentInfo methods. Reduce log noise in system journals

+++ segmentstore/storage/src/main/java/io/pravega/segmentstore/storage/chunklayer/TruncateOperation.java
@@ -83,6 +83,7 @@ class TruncateOperation implements Callable<CompletableFuture<Void>> {
 
                             if (segmentMetadata.getStartOffset() >= offset) {
                                 // Nothing to do
+                                logEnd();
                                 return CompletableFuture.completedFuture(null);
                             }
                             val oldChunkCount = segmentMetadata.getChunkCount();
@@ -113,7 +114,10 @@ class TruncateOperation implements Callable<CompletableFuture<Void>> {
                                                 }
 
                                                 // Remove read index block entries.
-                                                chunkedSegmentStorage.deleteBlockIndexEntriesForChunk(txn, streamSegmentName, oldStartOffset, segmentMetadata.getStartOffset());
+                                                // To avoid possibility of unintentional deadlock, skip this step for storage system segments.
+                                                if (!segmentMetadata.isStorageSystemSegment()) {
+                                                    chunkedSegmentStorage.deleteBlockIndexEntriesForChunk(txn, streamSegmentName, oldStartOffset, segmentMetadata.getStartOffset());
+                                                }
 
                                                 // Collect garbage.
                                                 return chunkedSegmentStorage.getGarbageCollector().addChunksToGarbage(txn.getVersion(), chunksToDelete)
